+++ src/main/java/com/Resimulators/simukraft/common/block/BlockConstructor.java
@@ -1,19 +1,15 @@
 package com.Resimulators.simukraft.common.block;
 
-import com.Resimulators.simukraft.client.interfaces.BaseJobGui;
-import net.minecraft.block.Block;
+import com.Resimulators.simukraft.client.gui.BaseJobGui;
 import net.minecraft.block.BlockState;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
 import net.minecraft.util.ActionResultType;
 import net.minecraft.util.Hand;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.BlockRayTraceResult;
 import net.minecraft.util.text.StringTextComponent;
 import net.minecraft.world.World;
-
-import javax.annotation.Nullable;
 
 public class BlockConstructor extends BlockBase {
     public BlockConstructor(final Properties properties,String name) {
